+++ SuperStar/HobbyPlugin/DxLib/DxLib/DXLibTime.h
@@ -10,7 +10,7 @@ namespace DXLib
     class Time final : public Platform::TimeInterface
     {
     public:
-        Uint32 VNowMSec() override final;
-        void VSleepMSec(const Uint32 in_uMs) override final;
+        Uint64 VNowMSec() override final;
+        void VSleepMSec(const Uint64 in_uMs) override final;
     };
 }  // namespace DXLib
